@@ -1,3 +1,3 @@
 package version
 
-var Version = "0.2.0"
+var Version = "0.3.0"
